+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Version 1.1.0 *(2013-12-24)*
 ----------------------------
 
  * DebugLn does not need a context in it's constructor.
+ * DebugLn uses thread name in message isntead of tag.
+
 
 Version 1.0.0 *(2013-12-21)*
 ----------------------------